LexAssist/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.0
requests==2.31.0
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq

sentence-transformers
faiss-cpu
pymupdfpip -m install
python-dotenv
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.3.0
requests==2.31.0
